separate pathbuilders and trajectorybuilder
